+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Fixed
 
 - `XIT ACT`: Fix the CX Buy action getting stuck when there are no orders in the order book and the "buy partial" toggle is on
+- `XIT ACT`: Fix opening a run tile for packages with configurable MTRA and no origins/destinations available
 
 ## 25.4.12